web/refactor: MainCardTrigger
